requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
django-extensions==3.2.3
frozenlist==1.5.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
openai==0.28.0
packaging==24.2
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dotenv==1.0.1
python-telegram-bot==21.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uWSGI==2.0.28
Werkzeug==3.1.3
yarl==1.18.3
